// src/types/response/category-management/mixer.ts
export interface IMixerManagement {
  mixer_id: string;
  _id: string;
  __v: number;
  batch_code_adx: string;
  batch_code_id: string;
  batch_name_adx: string;
  batch_name_id: string;
  createdAt: string;
  name: string;
  qc_trigger: string;
  qc_trigger_adx: string;
  real_ph_adx: any;
  sequence: number;
  start_trigger: string;
  start_trigger_adx: string;
  status: string;
  tag: string;
  updatedAt: string;
}